fix: format views and fix file handle error
