+++ views/navbar.ejs
@@ -1,9 +1,11 @@
 <nav class="navbar navbar-expand-sm navbar-expand-lg navbar-dark bg-dark">
-    <a class="navbar-brand" href="/">🛸</a>
-    <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-      <div class="navbar-nav">
-        <a class="nav-item nav-link active" href="/">UFO Sightings <span class="sr-only">(current)</span></a>
-        <a class="nav-item nav-link" href="/shapes">UFO Shapes</a>
-      </div>
+  <a class="navbar-brand" href="/">🛸</a>
+  <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+    <div class="navbar-nav">
+      <a class="nav-item nav-link active" href="/"
+        >UFO Sightings <span class="sr-only">(current)</span></a
+      >
+      <a class="nav-item nav-link" href="/shapes">UFO Shapes</a>
     </div>
+  </div>
 </nav>
