+++ views/singleSighting.ejs
@@ -1,5 +1,4 @@
-<%- include('header'); -%>
-<%- include('navbar'); -%>
+<%- include('header'); -%> <%- include('navbar'); -%>
 <body>
   <div class="container">
     <div class="row mt-3">
@@ -12,7 +11,7 @@
         </div>
         <div class="card-body">
           <h5 class="card-title">UFO Sighting <%= Number(index) + 1 %></h5>
-          <h6 class="card-subtitle mb-2 text-muted"> <%= sighting.city %></h6>
+          <h6 class="card-subtitle mb-2 text-muted"><%= sighting.city %></h6>
           <h6 class="card-subtitle mb-2 text-muted"><%= sighting.state %></h6>
           <p class="card-text"><%= sighting.text %></p>
         </div>
@@ -36,7 +35,7 @@
           </table>
         </div>
       </div>
-      <br>
+      <br />
     </div>
     <div class="row mt-3 justify-content-center">
       <a href="/" class="btn btn-primary">Back to all UFO Sightings</a>
@@ -44,6 +43,5 @@
   </div>
 
   <!-- <a href="/favorite?index=<%= index %>" class="btn btn-primary">Favorite</a> -->
-
 </body>
 <%- include('footer'); -%>
